% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descendants.R
\name{descendant}
\alias{descendant}
\title{Get a Node's Descendants}
\usage{
descendant(nodes, include = TRUE)
}
\arguments{
\item{nodes}{the node to get the descendants.}

\item{include}{should the descriptions be included? Default it TRUE.}
}
\description{
Get a Node's Descendants
}
