% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node.match.R
\name{node.match}
\alias{node.match}
\title{Match Nodes}
\usage{
node.match(nodes, nodelist, leaves, print.it = TRUE)
}
\arguments{
\item{nodes}{the nodes to match.}

\item{nodelist}{the list of nodes.}

\item{leaves}{the leaves to match to.}

\item{print.it}{should the result be printed? Default is \code{TRUE}.}
}
\description{
Match Nodes
}
