% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf-class-diag.R
\name{rf_class_error_diagnostics}
\alias{rf_class_error_diagnostics}
\title{Get the Error Diagnostics from an \code{rfsrc} Object}
\usage{
rf_class_error_diagnostics(x)
}
\arguments{
\item{x}{an \code{rfsrc} object.}
}
\value{
A tibble with a single row. Each column corresponds to an \code{rfsrc} diagnostic.
}
\description{
Get the Error Diagnostics from an \code{rfsrc} Object
}
